Add Project relationship to PageView
When a PageView is created, the tracking ID (tid) is used to
assign a project to the PageView. The tracking ID is evaluated
in the script and collect views to ensure validity before
assigning during the PageView creation.

A valid tid:
- Exists (not None)
- 'PA-' prefix
- 12 characters long
- Uppercase or Numbers
- A project currently exists with the tracking ID

If the tracking ID is determined to be invalid when the user is
requesting the script, a 404 will be raised to prevent a new request.

If the tracking ID is determined to be invalid when the user is
requesting the GIF, logging of analytics data is skipped but a 200
response is still returned.