Add view counts to admin site Project list display
Before this commit, there was no interface for users to see any
aggregation of the analytics being captured. View counts for the past
24 hours, 7 days, and 30 days were added as buckets on the admin page.
Total unique visits and the top path visited were also added to the
list display page.

The admin site is being used as the interface for now until a longer
term and more robust solution is decided.